// src/spacetime-bindings/types.ts
export const Game = __t.object("Game", {
  id: __t.u64(),
  schemaVersion: __t.u8(),
  tournamentId: __t.option(__t.u64()),
  startedAt: __t.timestamp(),
  completedAt: __t.option(__t.timestamp()),
  targetScore: __t.u16(),
  variant: __t.u8(),
  team1Seat0ModelId: __t.u32(),
  team1Seat2ModelId: __t.u32(),
  team2Seat1ModelId: __t.u32(),
  team2Seat3ModelId: __t.u32(),
  team1Score: __t.i16(),
  team2Score: __t.i16(),
  team1Bags: __t.u8(),
  team2Bags: __t.u8(),
  winnerTeam: __t.u8(),
  rngSeed: __t.u64(),
  allowRenege: __t.bool(),
  chatPolicy: __t.u8(),
  promptCheatingMode: __t.u8(),
  promptForDetection: __t.bool(),
  announceDetectedCheats: __t.bool(),
  agentDetectionQuorum: __t.bool(),
  cheatConsequenceKind: __t.u8(),
  cheatConsequenceValue: __t.i16(),
  spadesLeadPolicy: __t.u8(),
  minimumTeamBid: __t.u8(),
});
export type Game = __Infer<typeof Game>;
// ...
export const Model = __t.object("Model", {
  id: __t.u32(),
  name: __t.string(),
  kind: __t.u8(),
  version: __t.string(),
  introducedAt: __t.timestamp(),
});
export type Model = __Infer<typeof Model>;